Remove outdated FIXME in StorageManager https://bugs.webkit.org/show_bug.cgi?id=198049
Reviewed by Geoffrey Garen.

Previously StorageManager messages were dispatched and handled in the same thread, so we could perform a message
check. Now they are on different thread, and we would not be able to verify those parameters when dispatching
the message.

* NetworkProcess/WebStorage/StorageManager.cpp:
(WebKit::StorageManager::createLocalStorageMap):
(WebKit::StorageManager::createTransientLocalStorageMap):
(WebKit::StorageManager::createSessionStorageMap):
(WebKit::StorageManager::destroyStorageMap):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246011 268f45cc-cd09-0410-ab3c-d52691b4dbfc