Fix 64-bit vs 32-bit mismatch in TileController.cpp https://bugs.webkit.org/show_bug.cgi?id=198540 <rdar://problem/51410851>
Reviewed by Alex Christensen.

TileController::blankPixelCountForTiles calculates its result as a
uint64_t, but returns it as an unsigned. The former is a 64-bit value,
while the latter can be a 32-bit value on some platforms. This
mismatch can lead to a compile-time error. Fix this by explicitly
casting the 64-bit value to an "unsigned".

No new tests -- no new functionality.

* platform/graphics/ca/TileController.cpp:
(WebCore::TileController::blankPixelCountForTiles):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246091 268f45cc-cd09-0410-ab3c-d52691b4dbfc