resize-observer/element-leak.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=198666
Patch by Cathie Chen <cathiechen@igalia.com> on 2019-06-19
Reviewed by Frédéric Wang.

It takes a very long time to construct iframes which create and remove 1000 elements. This would cause timeout sometimes.
In order to make it more effective, reduce the number of elements to 200 and put them into a container first,
then attach the container to DOM tree.

* resize-observer/resources/element-leak-frame.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246584 268f45cc-cd09-0410-ab3c-d52691b4dbfc