Safari hanging while loading pages - WebCore::AccessibilityRenderObject::visiblePositionRangeForLine. https://bugs.webkit.org/show_bug.cgi?id=199434 <rdar://problem/52475140>
Patch by Andres Gonzalez <andresg_22@apple.com> on 2019-07-03
Reviewed by Chris Fleizach.

Source/WebCore:

Tests were disabled until underlying bug is fixed.

Rolling out a workaround for CharacterIterator::advance bug because it
seems to be causing a hang in Safari.
* editing/Editing.cpp:
(WebCore::visiblePositionForIndexUsingCharacterIterator):

LayoutTests:

Disabled the tests that were passing for workaround. The workaround had
to be roll out cause it is hanging Safari.
* TestExpectations:
* platform/ios-wk2/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247093 268f45cc-cd09-0410-ab3c-d52691b4dbfc