URLParser should reserialize IPv6 addresses that end with a colon https://bugs.webkit.org/show_bug.cgi?id=199627
Reviewed by Brady Eidson.

Source/WTF:

When an IPv6 address ends in a colon, the colon should be removed which means the
serialized output is different than the input, which the URLParser calls a syntax violation.
This matches the URL specification, and I added a unit test that used to assert.

* wtf/URLParser.cpp:
(WTF::URLParser::parseIPv6Host):

Tools:

* TestWebKitAPI/Tests/WTF/URLParser.cpp:
(TestWebKitAPI::TEST_F):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247267 268f45cc-cd09-0410-ab3c-d52691b4dbfc