REGRESSION: Chase app crashes immediately after authentication https://bugs.webkit.org/show_bug.cgi?id=199798 <rdar://problem/53106179>
Reviewed by Alex Christensen.

After r247411, we sometimes try to consult `_page` from WKWebView, before WKWebView's `_page` has been set. This
happens under a check for -[WKWebView _isEditable], from within -[WKContentView setupInteraction], which happens
if the host app is not linked on the iOS 12 SDK or earlier. To mitigate this, simply add a null check in
-_isEditable.

There's currently no great way to test this, since we don't have a mechanism for simulating linked-on checks.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _isEditable]):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247436 268f45cc-cd09-0410-ab3c-d52691b4dbfc