[WHLSL] The lexer should not choke on a single '/' https://bugs.webkit.org/show_bug.cgi?id=199886
Reviewed by Myles Maxfield.

Source/WebCore:

The bug is trivial: just recognize that a single '/' without a '*' or '/' afterwards is not the beginning of a comment.

While investigating this bug, I also found that part of the parser fails to properly propagate errors, so I fixed it.

New test in LayoutTests/webgpu/whlsl/float-math.html
I also added a new flag to WHLSLPrepare.cpp. It can be used to parse the full standard library (this is how I found this bug in the first place).

* Modules/webgpu/WHLSL/WHLSLLexer.cpp:
(WebCore::WHLSL::Lexer::skipWhitespaceAndComments):
* Modules/webgpu/WHLSL/WHLSLParser.cpp:
(WebCore::WHLSL::Parser::parsePossibleTernaryConditional):
* Modules/webgpu/WHLSL/WHLSLPrepare.cpp:
(WebCore::WHLSL::prepareShared):
* Modules/webgpu/WHLSL/WHLSLStandardLibraryUtilities.cpp:
(WebCore::WHLSL::includeStandardLibrary):
* Modules/webgpu/WHLSL/WHLSLStandardLibraryUtilities.h:

LayoutTests:

Added a new basic test of division.

* webgpu/whlsl/float-math.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247554 268f45cc-cd09-0410-ab3c-d52691b4dbfc