picture-in-picture.html fails because webkitpresentationmodechanged sometimes dispatched multiple times using Apple Internal build <rdar://problem/36455352>
Workaround by only listening for the first webkitpresentationmodechanged event dispatched. This test
is the canary in the coal mine that revealed that multiple webkitpresentationmodechanged events are
dispatched when one is expected. However this was not the primary purpose of the test and in absence
of a timeframe for a fix for <rdar://problem/36455352> work around this bug to avoid losing test coverage
when using an Apple Internal build.

* platform/ipad/media/controls/resources/picture-in-picture.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248011 268f45cc-cd09-0410-ab3c-d52691b4dbfc