KeyboardInputTests.CaretSelectionRectAfterRestoringFirstResponder API tests time out on iPad https://bugs.webkit.org/show_bug.cgi?id=200604 <rdar://problem/51273130>
Reviewed by Megan Gardner.

Tweak some API tests so that they work on iPad simulator. These tests checked that the final caret rect was
{{16, 13}, {2, 15}}; however, this is only correct behavior on iPhone, where we will scale the page so that the
focused element's font size is legible. Note that when the page is scaled, we scale the height but not the
width of the caret, which is why the width of the caret (in content coordinates) decreases while the height
remains the same.

We don't have the same behavior on iPad, so the expected caret rect is {{16, 13}, {3, 15}}, which is equal to
the caret rect at initial scale 1.

* TestWebKitAPI/Tests/ios/KeyboardInputTestsIOS.mm:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248496 268f45cc-cd09-0410-ab3c-d52691b4dbfc