Disabling text autosizing should prevent text autosizing https://bugs.webkit.org/show_bug.cgi?id=200955 <rdar://problem/48095374>
Reviewed by Simon Fraser.

Source/WebCore:

Before this change there was no way to reliably change the text size on iOS when a -webkit-text-size-adjust:none style was applied
because it would ignore the WKPreferences value for _textAutosizingEnabled.  Now that value has the desired effect when laying out.

* css/StyleResolver.cpp:
(WebCore::StyleResolver::checkForTextSizeAdjust):

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKitCocoa/TextSize.mm: Added.
(TEST):

LayoutTests:

* fast/text-autosizing/ios/text-size-adjust-inline-style.html:
The comment was incorrect because we call WKPreferencesSetTextAutosizingEnabled(preferences, false) in TestController::platformResetPreferencesToConsistentValues.
In order to test the effect of text autosizing in this test, we need to explicitly turn it on.  The disabling of text autosizing didn't do anything before this
patch, so it wasn't needed before now.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248966 268f45cc-cd09-0410-ab3c-d52691b4dbfc