Do not adjust viewport if editing selection is already visible https://bugs.webkit.org/show_bug.cgi?id=200907 <rdar://problem/53903417>
Reviewed by Simon Fraser.

Source/WebCore:

Test: fast/scrolling/ios/autoscroll-input-when-very-zoomed.html

Currently due to scrolling being mostly handled by integers, we are getting
issues with rounding errors when trying to adjust the viewport while
editing text when we are significantly zoomed in. The real fix would be to
start dealing with scrolling with floats/doubles, but until such time,
we should early out of adjusting selections that we are certain are currently
visible.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::scrollRectToVisible):

LayoutTests:

* fast/scrolling/ios/autoscroll-input-when-very-zoomed-expected.txt: Added.
* fast/scrolling/ios/autoscroll-input-when-very-zoomed.html: Added.
* resources/ui-helper.js:
(window.UIHelper.immediateZoomToScale):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248977 268f45cc-cd09-0410-ab3c-d52691b4dbfc